Finished frontend/backend

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,21 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 const WorkoutDetails = ( {workout} ) => {
-
+    const {user} = useAuthContext()
     const { dispatch } = useWorkoutsContext()
 
     const handleClick = async () =>
     {
+        if(!user)
+        {
+            return
+        }
         const response = await fetch('/api/workouts/' + workout._id,
             {
-                method: 'DELETE'
+                method: 'DELETE',
+                headers: {'Authorization': `Bearer ${user.token}`
+                }
             })
         const json = await response.json()
